+++ contracts/interfaces/randomness.sol
@@ -1,6 +1,7 @@
 pragma solidity ^0.6.2;
 
 interface RandomnessInterface {
-    function randomNumber(uint) external view returns (uint);
-    function getRandom(uint, uint) external;
+    function randomNumber(uint256) external view returns (uint256);
+
+    function getRandom(uint256, uint256) external;
 }
